Working pause menu

+++ Projectfiles/Level1.h
@@ -10,10 +10,11 @@
 @interface Level1 : CCLayer
 {
     int newScoreLevel1;
-    int shipSpeed;
+    float shipSpeed;
     CCLabelTTF * scoreLabel;
     CCAction *spinningCoin;
     NSMutableArray* spinningFrames;
+
     
 }
